refactor(type): 구입 금액 및 발행 개수 int로 변경
기존에는 로또 구입 금액 및 발행 개수 관련된 값은 모두 long으로
관리하였다. 그러나 해당 애플리케이션에서는 발행된 로또를 ArrayList에
저장하는데 이때 size(), add()등을 호출 시 내부 로직은 전부 int 타입으로
처리가 된다.
또한 현실적으로도 int타입의 범위를 넘어가는 금액으로 로또를 사는 경우가
거의 없기 때문에 해당 값을 모두 int 타입으로 변경했다.

+++ src/main/java/lotto/domain/service/LottoService.java
@@ -38,10 +38,10 @@ public class LottoService {
         winningDetails.put(FIFTH, 0);
     }
 
-    public void buyLotteries(long price) {
+    public void buyLotteries(int price) {
         validate(price);
-        long count = price / PRICE;
-        for (long i = 0; i < count; i++) {
+        int count = price / PRICE;
+        for (int i = 0; i < count; i++) {
             buyOneLotto();
         }
     }
@@ -100,18 +100,18 @@ public class LottoService {
 
     //TODO: validateMoreThanThousand 등을 인터페이스, 람다식 등을 이용해 변경
     //      그러면 하위 메서드를 계속 정의할 필요가 없을듯
-    private static void validate(long price) {
+    private static void validate(int price) {
         validateMoreThanThousand(price);
         validateMultipleOfThousand(price);
     }
 
-    private static void validateMultipleOfThousand(long price) {
+    private static void validateMultipleOfThousand(int price) {
         if (price % PRICE != 0) {
             throw new IllegalArgumentException(PRICE_NOT_DIVIDED_BY_THOUSAND);
         }
     }
 
-    private static void validateMoreThanThousand(long price) {
+    private static void validateMoreThanThousand(int price) {
         if (price < PRICE) {
             throw new IllegalArgumentException(PRICE_LESS_THAN_THOUSAND);
         }
